fix: issues with schema

+++ tools/build.sh
@@ -1,7 +1,7 @@
 index="src/index.ts"
 name="cli-thumus-app"
 builddir="./build"
-mkdir build
-rm $builddir/$name
-bun build $index --compile --outfile $builddir/$name
-# && upx -1 --lzma --overlay=skip $builddir/$name-tmp -o $builddir/$name # Doesn't work, but bun is very large so we want to compress it
+mkdir -p build
+[ -f "$builddir/$name" ] && rm $builddir/$name
+[ -f "$builddir/$name-tmp" ] && rm $builddir/$name-tmp
+bun build $index --compile --outfile $builddir/$name # -tmp && upx -1 --lzma --overlay=skip $builddir/$name-tmp -o $builddir/$name # Doesn't work, but bun is very large so we want to compress it
